backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-socketio==5.9.0
python-engineio==4.7.1
Pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
twain-wia-sane-scanner==1.0.0
pysane==2.9.4.5
pywin32==305
pyobjc==9.2
aiofiles==23.2.1
asyncio==3.4.3
